# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bitcoin AI Analyst Backend",
    "version": "2026-09-research-evidence-v2"
  },
  "paths": {
...
    "/api/volume-profile/current": {
      "get": {
        "tags": [
          "VolumeProfile"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "4h"
            }
          },
          {
            "name": "lookbackBars",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/volume-profile/calculate": {
      "post": {
        "tags": [
          "VolumeProfile"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "4h"
            }
          },
          {
            "name": "lookbackBars",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
...
}